// src/utils/types.ts
export type BannerPropsType = {
  data: {
    backgroundBannerImage?: string
    heading: string
    description?: string
    actionBtn1?: string
    actionBtn2?: string
  }
}